leetcode-java/src/main/java/lc239SlidingWindowMaximum/Solution.java
package lc239SlidingWindowMaximum;

import java.util.Deque;
import java.util.LinkedList;

/*
239. 滑动窗口最大值
给你一个整数数组 nums，有一个大小为 k 的滑动窗口从数组的最左侧移动到数组的最右侧。你只可以看到在滑动窗口内的 k 个数字。滑动窗口每次只向右移动一位。

返回滑动窗口中的最大值。



示例 1：

输入：nums = [1,3,-1,-3,5,3,6,7], k = 3
输出：[3,3,5,5,6,7]
解释：
滑动窗口的位置                最大值
---------------               -----
[1  3  -1] -3  5  3  6  7       3
 1 [3  -1  -3] 5  3  6  7       3
 1  3 [-1  -3  5] 3  6  7       5
 1  3  -1 [-3  5  3] 6  7       5
 1  3  -1  -3 [5  3  6] 7       6
 1  3  -1  -3  5 [3  6  7]      7
示例 2：

输入：nums = [1], k = 1
输出：[1]
示例 3：

输入：nums = [1,-1], k = 1
输出：[1,-1]
示例 4：

输入：nums = [9,11], k = 2
输出：[11]
示例 5：

输入：nums = [4,-2], k = 2
输出：[4]


提示：

1 <= nums.length <= 10^5
-10^4 <= nums[i] <= 10^4
1 <= k <= nums.length
 */
public class Solution {
    /**
     * 使用单调递减队列
     */
    public int[] maxSlidingWindow(int[] nums, int k) {
        int[] windows = new int[nums.length - k + 1];
        Deque<Integer> deque = new LinkedList<>();
        for (int i = 0; i < nums.length; i++) {
            // 保持队列的单调递减性
            while (!deque.isEmpty() && deque.peekLast() < nums[i]) {
                deque.pollLast();
            }
            deque.offerLast(nums[i]);

            // 队列元素不足 k 个，无法取得最大值
            if (i < k - 1) {
                continue;
            }

            // 队头元素就是窗口内最大得值
            windows[i - k + 1] = deque.getFirst();

            // 去除滑出窗口的元素
            if (deque.getFirst() == nums[i - k + 1]) {
                deque.pollFirst();
            }
        }

        return windows;
    }
}

